Moves model specific loading code to regressor scripts
Now that all others bugbug related scripts are using only the bugbug HTTP
service, moves the model loading code to the only scripts that needs it. We
will remove it in the future.
